requirements.txt
functions-framework==3.*
google-api-python-client
google-auth
google-auth-httplib2
requests
google-cloud-compute